DjangoProject/users/serializers.py
from rest_framework import serializers
from .models import User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username', 'email', 'role', 'first_name', 'last_name']
        # On peut aussi ajouter 'first_name' et 'last_name' car AbstractUser les possède déjà